JribbleFormatting prints references to packages with 'package' keyword.
With this change any reference to fully qualified name looks like this:
(package com.bar).Foo where Foo is a class in com.bar package.

This change to jribble syntax was needed to be able to distinguish between
two interpretation of the same expression: com.bar.Foo.

First valid interpretation of that expression is that we access local
variable (or field) named 'com', then we access field 'bar' and then we
access field 'Foo'.

Second interpretation would be that we access class named 'Foo' in package
'com.bar'.

We strive to make jribble rich enough so it can be parsed as locally as
possible thus we introduce this change.

Change-Id: I68db2404b74a0a725968aa5d00440bdb453055b4
Signed-off-by: Grzegorz Kossakowski <grek@google.com>