Fixed bug in handling of if expression in jribble target.
In case that at least one branch of if returns unboxed Unit if construct
must be rewritten into statement.

Added many tests which specify how if is being treated in different
situations. The most tricky part is about Unit type which can be
boxed or unboxed. It looks like it's being handled correctly in all
cases.

Change-Id: Ie1b522e121f20cf0507e1267a69587031cf4d03f
Signed-off-by: Grzegorz Kossakowski <grek@google.com>