Fixed handling of == operator in jribble.
Sala unifies use of == operator but there's no such unification in jribble.
Basically, jribble follows Java so Scala's == should be translated to:
  * == if both operands are subtypes of AnyVal (wich correspond to
    primitive types in jribble)
  * equals() call if both operands are subtypes of AnyRef

In order to take into account this rules tree rewrite case has been
introduced to NormalizeForJribble where we look for use of == and if
operands are subtypes of AnyRef we rewrite them to externalEquals call
which is Scala's implementation of equals that take care of handling nulls
and delegates to equals() defined on given operand for final judgement.

Added test which hopefully covers all important cases for use of ==
operator.

Change-Id: I5039230a946d39b02170f65d76af4d8c983e82fe
Signed-off-by: Grzegorz Kossakowski <grek@google.com>