Fixed many bugs related to handling traits/objects in jribble.
There were many problems with printing mirror class for objects. Firstly,
not in every case where we need to print mirror class we can use
companionSymbol because it might be NoSymbol. Secondly, we would try to
emit mirror class when there was no need for that. Fixed these problems
by following the implementation in GenJVM.

There are ClassDefs that don't extend any class and don't implements
any interfaces so changed the code to not assume that ClassDef have any
parents.

Check deciding if we print interface or class was buggy and has been
replaced by check based on Symbol's flags which should be more reliable.

Fixed referring to modules which was translated as reference to static
field $MODULE. There was an extra dollar character which is not needed
anymore. This change causes lots of changes to existing test cases which
were referring modules like Predef.

Fixed many bugs in printFlags method. Changes to this method was
inspired by implementation of similar method in GenJVM. More carefull
checking of printFlags method is still needed.

Added trait.scala test case which covers all fixed bugs by using mix
of nested traits and objects which have non-straightforward translation
to jribble interfaces/classes.

Change-Id: I486b18a97e517d64edfaf6b72fdbc16adace1b19
Signed-off-by: Grzegorz Kossakowski <grek@google.com>