Added support for patterns corresponding to Java switch statements.
Compiler will leave patterns involving Ints intact so they are
represented by match nodes. This gives an opportunity to map them into
Java switch statements that jribble supports.

Since switch is a statement in jribble we need to apply a transformation
that moves it into statement position. This is being done in
NormalizeForJribble. The code was inspired by similar transformation
performed for If node.

Added test-case for this kind of patterns assuring that the logic is
correct.

Change-Id: I2bfdd7fb49be56618b4943eb24be902762f76105
Signed-off-by: Grzegorz Kossakowski <grek@google.com>