If expressions are always translated to if statements in jribble.
It turned out that the code handling translation of if expressions
into conditional expressions had exponential complexity due to fact
that it would process recursively all child nodes in a tree for every
node separately.

By removing the check if we can translate given if expression into
conditional expression we get rid of exponential complexity. Obviously,
this is not a permanent solution but more a work-around.

Unfortunately, it looks like fixing this bug will require more involved
changes to NormalizeForJribble code.

Adapted test-cases to this change. Some test cases had to be change only
because numbering of local, synthetic variables change due to this commit.

Change-Id: Iac2567912b5c5ebf4591f7be00c2c11f797954c3
Signed-off-by: Grzegorz Kossakowski <grek@google.com>