Fixed bugs in static forwarders generation.
There were many issues with code generating static
forwarders to methods defined in objects. The most
prominent example is that static forwarders were
not generated for companion objects.

Now that code tries to follow corresponding code
from GenJVM that is a golden standard for this
piece of logic. Hopefully this fixes all bugs were
jribble was inconsistent with corresponding class
files.

Updated test-cases to reflect on this change.

Change-Id: I8e638b55cdab45438eaaf17329def6eb0a080ad4
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>