Remove hacks for forward jumps and admit we don't handle them.
Previous logic for handling some cases of forward jumps
was very, very fragile and was causing more harm than
good.

It turns out that handling of forward jumps is not very
critical because only a few classes in Scala library
depend on them.

Therefore, current strategy is to detect forward jumps
and throw away body of a method enclosing the jump
completely. We just replace it with an statement
throwing an exception. This will provide us a neccesary
info at the runtime just in case we've run into code
fragment that depends on forward jumps.

Fixed #7.

Change-Id: I313ac5772ffc9f3e0fb4846cab143b7948e57dfb
Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>