Initial checking of Miguel's jdk2ikvm plug-in.
Miguel Garcia created Scala compiler plug-in
that does source to source transformation that
we want to use to prepare custom version of
scala library adjusted to gwt needs and
limitations.

We are just copying sources of Miguel's
plug-in because it's not provided in any other
form for now. We might consider hosting this
plug-in on GitHub so improvement might be
shared between Miguel and our project but this
is not very important thing to do at the moment.

Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>