Added dependency on gwt when building scala-library-jribble.jar
Using method described in following thread:

https://groups.google.com/d/topic/scalagwt/AUp4r9Qu9Zg/discussion

we can reproduce environment that GWT provides. This makes it
easy to find out dependencies on Java classes that GWT doesn't
support.

This change is a quick hack until we find a better way to structure
our build. This will happen only when Scala migrates to SBT.

Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>