Fix problems with latest Scala IDE.
Latest version of Scala IDE will report error
if val is used in for-comprehension. We're fixing
this here.

The whole problem comes from the fact that we
are behind trunk (that includes those fixes) but
we must stay behind trunk because it has some
regressions that are breaking us.

Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>

Conflicts:

	src/compiler/scala/reflect/internal/Types.scala
	src/library/scala/Responder.scala