Implemented Factory Manifests that do not depend on reflection.
Added implementation of `factorymanifests` compiler phase.
It rewrites calls to ClassManifest.classType method with
calls to FactoryManifest.classType.

FactoryManifests require instance of Factory to be supplied
in order to create them. Compiler phase takes care of that
by generating anonymous classes implementing Factory trait.

Added this phase as part of Jribble platform. Factory
manifests is implemented as a plugin but we make it
required for jribble backend so people won't make a
mistake to not include it when compiling using jribble
backend.

Added a few tests covering this functionality.

Also, had to modify partest to include the plugin
when jribble tests are being executed.

Fixes #6.

Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>