Fix wrong object initialization order.
The exact problem is described in ticket #18.
In short, we should initialize MODULE$ field generated
for object as soon as superconstructor is called and
not wait until rest of constructor is executed.

Logic that implements this behaviour follows very
closely the one from jvm backend.

For things like that, it would be *far more* preferable
to share the logic between GenJVM and Jribble backend.
Ultimately, this should result in a separate compiler
phase (executed after cleanup and before icode) that
both jvm and jribble backends would share. However, this
is rather major refactoring and will require some time
to implement so I'm pushing this change for now.

Also, updated test-cases to reflect this change.

Fixes #18.

Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>