Build scala library for gwt with continuations enabled.
In order to do it cleanly we had to add support for
replacements for files not located in src/library
directory. Thus I refactored build.xml to first
assemble all scala files contributing to
scala-library-gwt.jar in one directory and only then
run jdk2ikvm tool on top of it.

While working on this I run into issue that simple
copying files from src/library to other directory
using Ant would make scala compiler to blow up
for some reason I don't understand. I decided to
go with a work-around by making a system call to
cp command that is capable of copying files in a way
that doesn't blow up scala compiler. This results
in platform-dependent build but I don't have enough
of time to fight this. Anyone interested in Windows
build can try to tackle this.

Added replacements for continuations library path that:

  * add work-arounds for pattern matcher shortcomings
  * replace NoSuchMethodException with RuntimeException

Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>