Fix bug in method signature emitting code.
Simple check if given symbol is constructor
is not enough. There are cases where constructors
are encoded as static methods. For example, trait
constructors are encoded like that.

We need to take that into account.

Created a test-case for this issue which shows
that signature with name "$init$" is emitted
instead of "new".

Signed-off-by: Grzegorz Kossakowski <grzegorz.kossakowski@gmail.com>