Explicitouter skips Matches when virtpatmat is enabled.
We disable transformation of Matches in ExplicitOuter
when Yvirtpatmat is enable and pattern can be translated
to switch table. In principle, when Yvirtpatmat is enabled
we should only see Matches that can be translated into
switch tables. Everything else should be rewritten in
typer. However, it turns out that Yvirtpatmat doesn't
handle (yet) patterns in try ... catch expressions so we
need to use old pattern matcher to handle them.

On the other hand, we don't want the old pattern matcher
to translate patterns that should be translated just to
switch statements - this is the job of backend. Especially
because old pattern matcher would introduce forward jumps
during the translation which is bad for Scala+GWT.

That's the rationale for a little bit dancing around patterns
until Yvirtpatmat handles pattern in try-catch expressions.