Enable Yvirtpatmat by default for Scala+GWT.
We should really have Yvirtpatmat enabled by
default and not bother our users with this
compiler option.

It's a bit hacky to enable it like that Scala
compiler has no API for doing it differently.
If this turns out to be problematic we might
need to introduce such an API.

Also, updated test-cases to reflect the
switch to Yvirtpatmat.